+++ src/main/java/at/hannibal2/skyhanni/features/nether/reputationhelper/FactionType.kt
@@ -7,6 +7,6 @@ enum class FactionType(val factionName: String, val apiName: String) {
 
     companion object {
         fun fromName(name: String) = entries.firstOrNull { it.factionName.equals(name, ignoreCase = true) }
-        fun fromAPIName(name: String) = entries.firstOrNull { it.apiName.equals(name, ignoreCase = true) }
+        fun fromAPIName(name: String): FactionType? = entries.firstOrNull { it.apiName.equals(name, ignoreCase = true) }
     }
 }
